refactor: move out debug map functions
