+++ sirc-vm/sbrc-vm/src/debug_adapter/mod.rs
@@ -1,2 +1,3 @@
+pub mod debug_map;
 pub mod server;
 pub mod types;
